requirements.txt
torch
torchvision
numpy
matplotlib
opencv-python
pillow
tqdm
tensorflow
mtcnn
flet
pandas